+++ src/gui/nmainmenubar.h
@@ -130,6 +130,7 @@ signals:
 
 public slots:
     void onOpenProjectWebPage();
+    void onOpenGettingStartedWebPage();
     void openThemeInformation();
 };